Rollback 0.4.0 instruction

+++ train.py
@@ -461,7 +461,7 @@ def evaluate(epoch=0):
 
         for idx, seg_iou in enumerate(eval_seg_iou_list):
             for jdx in range(len(score_thresh)):
-                seg_correct[idx, jdx] += (this_iou[jdx] >= seg_iou).float()
+                seg_correct[idx, jdx] += (this_iou[jdx] >= seg_iou)
 
         seg_total += 1
